# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Diwa",
        "description": "Diwa is an unofficial simple API from [distrowatch.com](https://distrowatch.com/)",
        "contact": {
            "email": "mzulfahmi807@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/Zzzul/diwa/blob/main/LICENSE"
        },
        "version": "2.0"
    },
    "paths": {
...
        "/api/v2/rankings/{slug}": {
            "get": {
                "tags": [
                    "Rankings"
                ],
                "summary": "Get top 100 distributions rankings with parameter",
                "description": "If {slug} not found, distrowatch.com will return the home page with default rankings(last 6 months). make sure {slug} is correct",
                "operationId": "findRankingnByParams",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Distribution Slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "trending-1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            }
        },
...
        "/api/v2/search": {
            "get": {
                "tags": [
                    "Distributions"
                ],
                "summary": "Get specific distribution",
                "description": "If one of the {params} not found/empty, distrowatch.com will used default params(All)",
                "operationId": "FilterDistribution",
                "parameters": [
                    {
                        "name": "ostype",
                        "in": "query",
                        "description": "OS Type",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Linux"
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Distribution Category",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "All"
                    },
                    {
                        "name": "origin",
                        "in": "query",
                        "description": "Country of Origin",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "All"
                    },
                    {
                        "name": "basedon",
                        "in": "query",
                        "description": "Based on",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Ubuntu"
                    },
                    {
                        "name": "notbasedon",
                        "in": "query",
                        "description": "Not Based on",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "None"
                    },
                    {
                        "name": "desktop",
                        "in": "query",
                        "description": "Desktop Interface",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Xfce"
                    },
                    {
                        "name": "architecture",
                        "in": "query",
                        "description": "Architecture",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "All"
                    },
                    {
                        "name": "package",
                        "in": "query",
                        "description": "Package Management",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "All"
                    },
                    {
                        "name": "rolling",
                        "in": "query",
                        "description": "Release Model",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "All"
                    },
                    {
                        "name": "isosize",
                        "in": "query",
                        "description": "Install Media Size",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "All"
                    },
                    {
                        "name": "netinstall",
                        "in": "query",
                        "description": "Install Method",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "All"
                    },
                    {
                        "name": "language",
                        "in": "query",
                        "description": "Multi Language Support",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "All"
                    },
                    {
                        "name": "defaultinit",
                        "in": "query",
                        "description": "Init Sofrware",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "All"
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Active"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            }
        },
...
        "/api/v2/weekly/{id}": {
            "get": {
                "tags": [
                    "Weekly News"
                ],
                "summary": "Get weekly news information detail",
                "description": "If {weekly_id} not found, distrowatch.com will return the latest weekly news. make sure {weekly_id} is correct",
                "operationId": "getWeeklyNewsById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Weekly News Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "20220502"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            }
        },
...
    },
...
}